# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Study Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ───── TOPICS ─────
  /subjects/{subjectId}/topics:
    get:
      operationId: listTopics
      tags: [topics]
      summary: List topics for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
    post:
      operationId: createTopic
      tags: [topics]
      summary: Create a topic
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{id}:
    get:
      operationId: getTopic
      tags: [topics]
      summary: Get a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
    patch:
      operationId: updateTopic
      tags: [topics]
      summary: Update a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
    delete:
      operationId: deleteTopic
      tags: [topics]
      summary: Delete a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ───── NOTES ─────
  /topics/{topicId}/notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes for a topic
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ───── FORMULAS ─────
...
  /quick-notes:
    get:
      operationId: listQuickNotes
      tags: [quicknotes]
      summary: List quick notes
      responses:
        "200":
          description: List of quick notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuickNote"
    post:
      operationId: createQuickNote
      tags: [quicknotes]
      summary: Create a quick note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuickNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickNote"
  /quick-notes/{id}:
    patch:
      operationId: updateQuickNote
      tags: [quicknotes]
      summary: Update a quick note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuickNoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickNote"
    delete:
      operationId: deleteQuickNote
      tags: [quicknotes]
      summary: Delete a quick note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ───── TODOS ─────
  /todos:
    get:
      operationId: listTodos
      tags: [todos]
      summary: List todos
      responses:
        "200":
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      operationId: createTodo
      tags: [todos]
      summary: Create a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
  /todos/{id}:
    patch:
      operationId: updateTodo
      tags: [todos]
      summary: Update a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
    delete:
      operationId: deleteTodo
      tags: [todos]
      summary: Delete a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /todos/{id}/complete:
    patch:
      operationId: completeTodo
      tags: [todos]
      summary: Toggle todo completion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"

  # ───── BOOKMARKS ─────
...
components:
  schemas:
...
    Subject:
      type: object
      required: [id, name, color, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        createdAt:
          type: string
          format: date-time
    SubjectInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
    SubjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string

    # ── Topic ──
    Topic:
      type: object
      required: [id, subjectId, name, createdAt]
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TopicInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    TopicUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]

    # ── Note ──
    Note:
      type: object
      required: [id, topicId, title, createdAt]
      properties:
        id:
          type: integer
        topicId:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NoteInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
    NoteUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: ["string", "null"]

    # ── Formula ──
...
    QuickNote:
      type: object
      required: [id, title, color, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QuickNoteInput:
      type: object
      required: [title, color]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        color:
          type: string
    QuickNoteUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: ["string", "null"]
        color:
          type: string

    # ── Todo ──
    Todo:
      type: object
      required: [id, text, completed, createdAt]
      properties:
        id:
          type: integer
        text:
          type: string
        completed:
          type: boolean
        subjectId:
          type: ["integer", "null"]
        subjectName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    TodoInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
        subjectId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
    TodoUpdate:
      type: object
      properties:
        text:
          type: string
        subjectId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        completed:
          type: boolean

    # ── Bookmark ──
...